requirements.txt
python-docx==0.8.11
weasyprint==60.2
reportlab==3.6.13
